remove ratio option since it prevented bookmarks from updating correctly
Basically, 200 bookmarks would be retrieved from Pinboard, but more would be in the local database. When using Set A as the database and the 200 as Set B, using A-B would cause local bookmarks to be deleted when they weren't actually deleted in Pinboard.

The best fix was to just pull the entire bookmark database for each request.